fix: split query handler
