+++ contracts/cosmwasm-vm/cw-icon-light-client/src/lib.rs
@@ -1,12 +1,13 @@
 pub use cw_light_client_common::constants;
+pub mod context;
 pub mod contract;
+pub mod query_handler;
 
 pub use cw_light_client_common::light_client;
 
 #[cfg(feature = "mock")]
 pub mod mock_client;
-pub use cw_light_client_common::query_handler;
-pub use cw_light_client_common::state;
+
 pub use cw_light_client_common::traits;
 
 pub use cw_light_client_common::error::ContractError;
