+++ contracts/cosmwasm-vm/cw-light-client-common/src/lib.rs
@@ -1,7 +1,5 @@
 pub mod constants;
 pub mod error;
 pub mod light_client;
-pub mod query_handler;
-pub mod state;
 pub mod traits;
 pub use crate::error::ContractError;
