+++ contracts/cosmwasm-vm/cw-light-client-common/src/light_client.rs
@@ -1,5 +1,6 @@
 use crate::constants::TRUST_LEVEL;
 
+use crate::traits::IQueryHandler;
 use crate::traits::{ConsensusStateUpdate, IContext, ILightClient};
 use crate::ContractError;
 use common::icon::icon::lightclient::v1::ConsensusState;
