+++ contracts/cosmwasm-vm/cw-wasm-light-client/src/lib.rs
@@ -2,10 +2,10 @@ pub use cw_light_client_common::constants;
 mod context;
 pub mod contract;
 mod msg;
-
+pub mod query_handler;
+pub mod utils;
 pub use cw_light_client_common::light_client;
-pub use cw_light_client_common::query_handler;
-pub use cw_light_client_common::state;
+
 pub use cw_light_client_common::traits;
 
 pub use cw_light_client_common::error::ContractError;
